fix(tldraw): display comma keyboard shortcuts
