+++ packages/tldraw/src/lib/ui/kbd-utils.ts
@@ -1,5 +1,14 @@
 import { tlenv } from '@tldraw/editor'
 
+/*!
+ * `splitKbd` is adapted from hotkeys-js.
+ *
+ * MIT License: https://github.com/jaywcjlove/hotkeys-js/blob/master/LICENSE
+ * Copyright (c) 2015-present, Kenny Wong
+ * Copyright (c) 2011-2013 Thomas Fuchs (https://github.com/madrobby/keymaster)
+ * Source: https://github.com/jaywcjlove/hotkeys-js
+ */
+
 // N.B. We rework these Windows placeholders down below.
 const cmdKey = tlenv.isDarwin ? '⌘' : '__CTRL__'
 const ctrlKey = tlenv.isDarwin ? '⌃' : '__CTRL__'
@@ -7,11 +16,8 @@ const altKey = tlenv.isDarwin ? '⌥' : '__ALT__'
 
 /** @public */
 export function kbd(str: string) {
-	if (str === ',') return [',']
-
 	return (
-		str
-			.split(',')[0]
+		(splitKbd(str)[0] ?? '')
 			// If the string contains [[Tab]], we don't split these up
 			// as they're meant to be atomic.
 			.split(/(\[\[[^\]]+\]\])/g)
@@ -45,6 +51,24 @@ export function kbd(str: string) {
 			})
 			.flat()
 	)
+}
+
+// Split a kbd string on commas, treating an empty entry produced by "x,," as a literal
+// trailing comma on the previous entry.
+/** @internal */
+export function splitKbd(key: string) {
+	if (!key) return []
+	key = key.replace(/\s/g, '')
+	const keys = key.split(',')
+	let index = keys.lastIndexOf('')
+
+	for (; index >= 0; ) {
+		keys[index - 1] += ','
+		keys.splice(index, 1)
+		index = keys.lastIndexOf('')
+	}
+
+	return keys
 }
 
 /** @public */
